requirements.txt
django==5.0
celery==5.3.6
redis==5.0.1
flower==2.0.1
psycopg2-binary==2.9.9
djangorestframework
djangorestframework-simplejwt
channels[daphne]
channels-redis
pandas
django-cors-headers
